Update LunaHsmManager to handle bad slot names
The LunaSlotManager treats a bad slot name differently from a bad
password: it does not throw an exception but quietly fails to log in.
This was revealed in unit testing.

To handle this behavior, we make two changes:
- Only call loadKeyStore() if we are definitely logged in
- Throw an IllegalArgumentException if are not logged in

Note that we also eliminate the 'logged in' check in loadKeyStore(),
since it is now a private method that should only be called when we
are in the correct state.